split up fussballde.py
